Move Mailboxes, Folders & Threads from MailData to MainViewModel
