+++ app/src/main/java/com/infomaniak/mail/ui/main/menu/FoldersAdapter.kt
@@ -23,9 +23,9 @@ import androidx.annotation.DrawableRes
 import androidx.appcompat.content.res.AppCompatResources
 import androidx.recyclerview.widget.RecyclerView
 import com.infomaniak.mail.R
-import com.infomaniak.mail.data.MailData
 import com.infomaniak.mail.data.models.Folder
 import com.infomaniak.mail.databinding.ItemFolderMenuDrawerBinding
+import com.infomaniak.mail.ui.main.MainViewModel
 import com.infomaniak.mail.ui.main.menu.FoldersAdapter.FolderViewHolder
 import com.infomaniak.mail.utils.context
 import com.infomaniak.lib.core.R as RCore
@@ -42,7 +42,7 @@ class FoldersAdapter(
     override fun onBindViewHolder(holder: FolderViewHolder, position: Int, payloads: MutableList<Any>) {
         if (payloads.firstOrNull() == Unit) {
             val folder = folders[position]
-            holder.binding.item.setSelectedState(MailData.currentFolderFlow.value?.id == folder.id)
+            holder.binding.item.setSelectedState(MainViewModel.currentFolderFlow.value?.id == folder.id)
         } else {
             super.onBindViewHolder(holder, position, payloads)
         }
@@ -77,7 +77,7 @@ class FoldersAdapter(
         indent = context.resources.getDimension(RCore.dimen.marginStandard).toInt() * (folderIndent ?: 0)
         badge = badgeText ?: ""
 
-        setSelectedState(MailData.currentFolderFlow.value?.id == id)
+        setSelectedState(MainViewModel.currentFolderFlow.value?.id == id)
 
         setOnClickListener { openFolder.invoke(id) }
     }
