+++ app/src/main/java/com/infomaniak/mail/ui/main/menu/user/SwitchUserFragment.kt
@@ -24,12 +24,13 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
-import com.infomaniak.mail.data.MailData
 import com.infomaniak.mail.databinding.FragmentSwitchUserBinding
 import com.infomaniak.mail.ui.LoginActivity
+import com.infomaniak.mail.ui.main.MainViewModel
 import com.infomaniak.mail.ui.main.menu.user.SwitchUserAccountsAdapter.UiAccount
 import com.infomaniak.mail.utils.AccountUtils
 import com.infomaniak.mail.utils.observeNotNull
@@ -38,20 +39,21 @@ import kotlinx.coroutines.launch
 
 class SwitchUserFragment : Fragment() {
 
+    private val mainViewModel: MainViewModel by activityViewModels()
     private val viewModel: SwitchUserViewModel by viewModels()
 
     private lateinit var binding: FragmentSwitchUserBinding
 
     private val accountsAdapter = SwitchUserAccountsAdapter { selectedMailbox ->
         if (selectedMailbox.userId == AccountUtils.currentUserId) {
-            MailData.selectMailbox(selectedMailbox)
+            MainViewModel.selectMailbox(selectedMailbox)
             findNavController().popBackStack()
         } else {
             lifecycleScope.launch {
                 AccountUtils.currentUser = AccountUtils.getUserById(selectedMailbox.userId)
                 AccountUtils.currentMailboxId = selectedMailbox.mailboxId
 
-                MailData.close()
+                mainViewModel.close()
 
                 AccountUtils.reloadApp?.invoke(bundleOf())
             }
