+++ app/src/main/java/com/infomaniak/mail/ui/main/newmessage/NewMessageViewModel.kt
@@ -20,12 +20,15 @@ package com.infomaniak.mail.ui.main.newmessage
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.infomaniak.mail.data.MailData
 import com.infomaniak.mail.data.api.ApiRepository
+import com.infomaniak.mail.data.cache.mailboxInfos.MailboxController
 import com.infomaniak.mail.data.cache.userInfos.ContactController
 import com.infomaniak.mail.data.models.Draft
+import com.infomaniak.mail.data.models.Mailbox
 import com.infomaniak.mail.data.models.Recipient
+import com.infomaniak.mail.ui.main.MainViewModel
 import com.infomaniak.mail.ui.main.newmessage.NewMessageActivity.EditorAction
+import com.infomaniak.mail.utils.AccountUtils
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.ext.toRealmList
 import io.realm.kotlin.types.RealmList
@@ -33,10 +36,14 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 class NewMessageViewModel : ViewModel() {
+
     val allContacts = MutableLiveData<List<UiContact>?>()
     val recipients = mutableListOf<UiContact>()
     val newMessageCc = mutableListOf<UiContact>()
     val newMessageBcc = mutableListOf<UiContact>()
+
+    val mailboxes = MutableLiveData<List<Mailbox>?>()
+
     var areAdvancedFieldsOpened = false
     var isEditorExpanded = false
     val editorAction = MutableLiveData<EditorAction>()
@@ -51,8 +58,14 @@ class NewMessageViewModel : ViewModel() {
         }
     }
 
+    fun listenToMailboxes() = viewModelScope.launch {
+        MailboxController.getMailboxesAsync(AccountUtils.currentUserId).collect {
+            mailboxes.value = it.list
+        }
+    }
+
     fun sendMail(draft: Draft, action: Draft.DraftAction) {
-        val mailbox = MailData.currentMailboxFlow.value ?: return
+        val mailbox = MainViewModel.currentMailboxFlow.value ?: return
         fun sendDraft() = ApiRepository.sendDraft(mailbox.uuid, draft.fillForApi("send"))
         fun saveDraft() = ApiRepository.saveDraft(mailbox.uuid, draft.fillForApi("save"))
